Defer initialization and add retry logic
zha initializes zha_gateway in async_setup_entry(), which may not have
happened at the time zha_map's async_setup() is run.

This moves all of the setup code to run after
EVENT_HOMEASSISTANT_START, and also implements a naive retry
loop. Retrying for a hardcoded length of time and then giving up only
makes this less likely to lose the race, but doesn't guarantee it.
A better implementation would perhaps be to trigger setup of zha_map
on the completion of zha's configuration.